backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
defusedxml==0.8.0rc2
Django==5.1.3
django-autoslug==1.9.9
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
fonttools==4.55.2
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
